[Porting one bug already fixed in trunk]
Bug fixed.

The Portland compiler did not like very long lines even if they are
just comments, and did not like to mix, in an interface block, the
function or subroutine interface definitions, with the module
procedure statements. I have looked up the standard, and I believe
both things are legal, but in any case this way that compiler is happy
and it does not harm the others compilers.