Backport of 15176 to branches/5.0.x: ------------------------------------------------------------------------ r15176 | dstrubbe | 2016-03-04 14:07:03 -0500 (Fri, 04 Mar 2016) | 2 lines
Bugfix: the nfft m4 script unwisely appends its CFLAGS and LIBS to the global ones (which should not be done regardless), even if it found that they did not work.

------------------------------------------------------------------------
Backport of 15178 to branches/5.0.x:
------------------------------------------------------------------------
r15178 | dstrubbe | 2016-03-04 15:34:48 -0500 (Fri, 04 Mar 2016) | 2 lines

Bugfix to r15176: while nothing should be appended to CFLAGS and LIBS in the nff4 m4 script, they should be restored to the saved values.

------------------------------------------------------------------------