Make all functions in the message module use intent(in)
Except for init and end which are supposed to change the object.

This is done in order to be able to call these functions from places
where the message object might be intent(in).

Where a write statement to a message object with intent(in) occures, a
local object is created with allocate(that, source=this). Then the write
statement can be used with this object while preserving the behavior of
the original object. Moreover, the memory overhead should be small,
because only 5 pointers are copied.