Make the lines buffer global and call it messages
This allows to use the messages functions from everywhere, also from
functions where the messages object is part of an intent(in) argument.
For these arguments, the compiler does not allow their members to appear
in write statements although the corresponding component is a pointer.

This solution now combines a global variable as buffer with local
objects that modify the behavior of how the messages are printed out.